+++ extensions/Worker.Extensions.EventHubs/src/EventDataConverter.cs
@@ -1,22 +1,22 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 using System;
 using System.Linq;
 using System.Threading.Tasks;
 using Azure.Core.Amqp;
-using Azure.Messaging.ServiceBus;
 using Microsoft.Azure.Functions.Worker.Converters;
 using Microsoft.Azure.Functions.Worker.Core;
 using Microsoft.Azure.Functions.Worker.Extensions.Abstractions;
-using Microsoft.Azure.Functions.Worker.Extensions.ServiceBus;
+using Azure.Messaging.EventHubs;
+using Microsoft.Azure.Functions.Worker.Extensions.EventHubs;
 
 namespace Microsoft.Azure.Functions.Worker
 {
     [SupportsDeferredBinding]
-    [SupportedConverterType(typeof(ServiceBusReceivedMessage))]
-    [SupportedConverterType(typeof(ServiceBusReceivedMessage[]))]
-    internal class ServiceBusReceivedMessageConverter : IInputConverter
+    [SupportedConverterType(typeof(EventData))]
+    [SupportedConverterType(typeof(EventData[]))]
+    internal class EventDataConverter : IInputConverter
     {
         public ValueTask<ConversionResult> ConvertAsync(ConverterContext context)
         {
@@ -24,10 +24,10 @@ namespace Microsoft.Azure.Functions.Worker
             {
                 ConversionResult result = context?.Source switch
                 {
-                    ModelBindingData binding => ConversionResult.Success(ConvertToServiceBusReceivedMessage(binding)),
+                    ModelBindingData binding => ConversionResult.Success(ConvertToEventData(binding)),
                     // Only array collections are currently supported, which matches the behavior of the in-proc extension.
                     CollectionModelBindingData collection => ConversionResult.Success(collection.ModelBindingDataArray
-                        .Select(ConvertToServiceBusReceivedMessage).ToArray()),
+                        .Select(ConvertToEventData).ToArray()),
                     _ => ConversionResult.Unhandled()
                 };
                 return new ValueTask<ConversionResult>(result);
@@ -38,7 +38,7 @@ namespace Microsoft.Azure.Functions.Worker
             }
         }
 
-        private ServiceBusReceivedMessage ConvertToServiceBusReceivedMessage(ModelBindingData binding)
+        private EventData ConvertToEventData(ModelBindingData binding)
         {
             if (binding?.Source is not Constants.BindingSource)
             {
@@ -52,14 +52,7 @@ namespace Microsoft.Azure.Functions.Worker
                     $"Unexpected content-type. Only '{Constants.BinaryContentType}' is supported.");
             }
 
-            // The lock token is a 16 byte GUID
-            const int lockTokenLength = 16;
-
-            ReadOnlyMemory<byte> bytes = binding.Content.ToMemory();
-            ReadOnlyMemory<byte> lockTokenBytes = bytes.Slice(0, lockTokenLength);
-            ReadOnlyMemory<byte> messageBytes = bytes.Slice(lockTokenLength, bytes.Length - lockTokenLength);
-            return ServiceBusReceivedMessage.FromAmqpMessage(AmqpAnnotatedMessage.FromBytes(BinaryData.FromBytes(messageBytes)),
-                BinaryData.FromBytes(lockTokenBytes));
+            return new EventData(AmqpAnnotatedMessage.FromBytes(binding.Content));
         }
     }
 }