+++ sdk/Sdk.Analyzers/Extensions/TypeSymbolExtensions.cs
@@ -1,11 +1,11 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
+using System.Text.RegularExpressions;
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Diagnostics;
 
 namespace Microsoft.Azure.Functions.Worker.Sdk.Analyzers
 {
@@ -37,6 +37,26 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Analyzers
             }
 
             return false;
+        }
+
+        internal static List<AttributeData> GetInputConverterAttributes(this ITypeSymbol attributeType, Compilation compilation)
+        {
+            var inputConverterAttributeType = compilation.GetTypeByMetadataName(Constants.Types.InputConverterAttribute);
+            return attributeType.GetAttributes()
+                .Where(attr => SymbolEqualityComparer.Default.Equals(attr.AttributeClass, inputConverterAttributeType))
+                .ToList();
+        }
+
+        internal static string GetMinimalDisplayName(this ITypeSymbol type, SemanticModel semanticModel)
+        {
+            string name = type.ToMinimalDisplayString(semanticModel, 0);
+
+            if (name.Contains("IEnumerable"))
+            {
+                name = Regex.Match(name, @"IEnumerable<[^>]+>").Value;
+            }
+
+            return name;
         }
     }
 }