+++ test/WorkerExtensionTests/EventHubs/EventDataConverterTests.cs
@@ -3,53 +3,51 @@
 
 using System;
 using System.Threading.Tasks;
-using Azure.Messaging.ServiceBus;
+using Azure.Messaging.EventHubs;
 using Google.Protobuf;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Converters;
-using Microsoft.Azure.Functions.Worker.Extensions.ServiceBus;
+using Microsoft.Azure.Functions.Worker.Extensions.EventHubs;
 using Microsoft.Azure.Functions.Worker.Grpc.Messages;
 using Microsoft.Azure.Functions.Worker.Tests.Converters;
 using Xunit;
 
 namespace Microsoft.Azure.Functions.WorkerExtension.Tests
 {
-    public class ServiceBusReceivedMessageConverterTests
+    public class EventDataConverterTests
     {
         [Fact]
         public async Task ConvertAsync_ReturnsSuccess()
         {
-            var lockToken = Guid.NewGuid();
-            var message = CreateReceivedMessage(lockToken);
+            var eventData = CreateEventData();
 
             var data = new GrpcModelBindingData(new ModelBindingData()
             {
                 Version = "1.0",
-                Source = "AzureServiceBusReceivedMessage",
-                Content = ByteString.CopyFrom(ConvertReceivedMessageToBinaryData(message)),
+                Source = "AzureEventHubsEventData",
+                Content = ByteString.CopyFrom(ConvertEventDataToBinaryData(eventData)),
                 ContentType = Constants.BinaryContentType
             });
             var context = new TestConverterContext(typeof(string), data);
-            var converter = new ServiceBusReceivedMessageConverter();
+            var converter = new EventDataConverter();
             var result = await converter.ConvertAsync(context);
 
             Assert.Equal(ConversionStatus.Succeeded, result.Status);
-            var output = result.Value as ServiceBusReceivedMessage;
+            var output = result.Value as EventData;
             Assert.NotNull(output);
-            AssertReceivedMessage(output, lockToken);
+            AssertEventData(output);
         }
 
         [Fact]
         public async Task ConvertAsync_Batch_ReturnsSuccess()
         {
-            var lockToken = Guid.NewGuid();
-            var message = CreateReceivedMessage(lockToken);
+            var message = CreateEventData();
 
             var data = new ModelBindingData
             {
                 Version = "1.0",
-                Source = "AzureServiceBusReceivedMessage",
-                Content = ByteString.CopyFrom(ConvertReceivedMessageToBinaryData(message)),
+                Source = "AzureEventHubsEventData",
+                Content = ByteString.CopyFrom(ConvertEventDataToBinaryData(message)),
                 ContentType = Constants.BinaryContentType
             };
 
@@ -58,36 +56,35 @@ namespace Microsoft.Azure.Functions.WorkerExtension.Tests
             array.ModelBindingData.Add(data);
 
             var context = new TestConverterContext(typeof(string), new GrpcCollectionModelBindingData(array));
-            var converter = new ServiceBusReceivedMessageConverter();
+            var converter = new EventDataConverter();
             var result = await converter.ConvertAsync(context);
 
             Assert.Equal(ConversionStatus.Succeeded, result.Status);
-            var output = result.Value as ServiceBusReceivedMessage[];
+            var output = result.Value as EventData[];
             Assert.NotNull(output);
             Assert.Equal(2, output.Length);
-            AssertReceivedMessage(output[0], lockToken);
-            AssertReceivedMessage(output[1], lockToken);
+            AssertEventData(output[0]);
+            AssertEventData(output[1]);
         }
 
         [Fact]
         public async Task ConvertAsync_ReturnsFailure_WrongContentType()
         {
-            var lockToken = Guid.NewGuid();
-            var message = CreateReceivedMessage(lockToken);
+            var eventData = CreateEventData();
 
             var data = new GrpcModelBindingData(new ModelBindingData()
             {
                 Version = "1.0",
                 Source = Constants.BindingSource,
-                Content = ByteString.CopyFrom(ConvertReceivedMessageToBinaryData(message)),
+                Content = ByteString.CopyFrom(ConvertEventDataToBinaryData(eventData)),
                 ContentType = "application/json"
             });
             var context = new TestConverterContext(typeof(string), data);
-            var converter = new ServiceBusReceivedMessageConverter();
+            var converter = new EventDataConverter();
             var result = await converter.ConvertAsync(context);
 
             Assert.Equal(ConversionStatus.Failed, result.Status);
-            var output = result.Value as ServiceBusReceivedMessage;
+            var output = result.Value as EventData;
             Assert.Null(output);
             Assert.IsType<InvalidOperationException>(result.Error);
         }
@@ -95,14 +92,13 @@ namespace Microsoft.Azure.Functions.WorkerExtension.Tests
         [Fact]
         public async Task ConvertAsync_Batch_ReturnsFailure_WrongContentType()
         {
-            var lockToken = Guid.NewGuid();
-            var message = CreateReceivedMessage(lockToken);
+            var message = CreateEventData();
 
             var data = new ModelBindingData
             {
                 Version = "1.0",
                 Source = Constants.BindingSource,
-                Content = ByteString.CopyFrom(ConvertReceivedMessageToBinaryData(message)),
+                Content = ByteString.CopyFrom(ConvertEventDataToBinaryData(message)),
                 ContentType = "application/json"
             };
 
@@ -111,11 +107,11 @@ namespace Microsoft.Azure.Functions.WorkerExtension.Tests
             array.ModelBindingData.Add(data);
 
             var context = new TestConverterContext(typeof(string), new GrpcCollectionModelBindingData(array));
-            var converter = new ServiceBusReceivedMessageConverter();
+            var converter = new EventDataConverter();
             var result = await converter.ConvertAsync(context);
 
             Assert.Equal(ConversionStatus.Failed, result.Status);
-            var output = result.Value as ServiceBusReceivedMessage[];
+            var output = result.Value as EventData[];
             Assert.Null(output);
             Assert.IsType<InvalidOperationException>(result.Error);
         }
@@ -123,22 +119,21 @@ namespace Microsoft.Azure.Functions.WorkerExtension.Tests
         [Fact]
         public async Task ConvertAsync_ReturnsFailure_WrongSource()
         {
-            var lockToken = Guid.NewGuid();
-            var message = CreateReceivedMessage(lockToken);
+            var eventData = CreateEventData();
 
             var data = new GrpcModelBindingData(new ModelBindingData()
             {
                 Version = "1.0",
                 Source = "some-other-source",
-                Content = ByteString.CopyFrom(ConvertReceivedMessageToBinaryData(message)),
+                Content = ByteString.CopyFrom(ConvertEventDataToBinaryData(eventData)),
                 ContentType = Constants.BinaryContentType
             });
             var context = new TestConverterContext(typeof(string), data);
-            var converter = new ServiceBusReceivedMessageConverter();
+            var converter = new EventDataConverter();
             var result = await converter.ConvertAsync(context);
 
             Assert.Equal(ConversionStatus.Failed, result.Status);
-            var output = result.Value as ServiceBusReceivedMessage;
+            var output = result.Value as EventData;
             Assert.Null(output);
             Assert.IsType<InvalidOperationException>(result.Error);
         }
@@ -146,14 +141,13 @@ namespace Microsoft.Azure.Functions.WorkerExtension.Tests
         [Fact]
         public async Task ConvertAsync_Batch_ReturnsFailure_WrongSource()
         {
-            var lockToken = Guid.NewGuid();
-            var message = CreateReceivedMessage(lockToken);
+            var message = CreateEventData();
 
             var data = new ModelBindingData
             {
                 Version = "1.0",
                 Source = "some-other-source",
-                Content = ByteString.CopyFrom(ConvertReceivedMessageToBinaryData(message)),
+                Content = ByteString.CopyFrom(ConvertEventDataToBinaryData(message)),
                 ContentType = Constants.BinaryContentType
             };
 
@@ -162,70 +156,36 @@ namespace Microsoft.Azure.Functions.WorkerExtension.Tests
             array.ModelBindingData.Add(data);
 
             var context = new TestConverterContext(typeof(string), new GrpcCollectionModelBindingData(array));
-            var converter = new ServiceBusReceivedMessageConverter();
+            var converter = new EventDataConverter();
             var result = await converter.ConvertAsync(context);
 
             Assert.Equal(ConversionStatus.Failed, result.Status);
-            var output = result.Value as ServiceBusReceivedMessage[];
+            var output = result.Value as EventData[];
             Assert.Null(output);
             Assert.IsType<InvalidOperationException>(result.Error);
         }
 
-        private static void AssertReceivedMessage(ServiceBusReceivedMessage output, Guid lockToken)
+        private static void AssertEventData(EventData output)
         {
-            Assert.Equal("body", output.Body.ToString());
+            Assert.Equal("body", output.EventBody.ToString());
             Assert.Equal("messageId", output.MessageId);
             Assert.Equal("correlationId", output.CorrelationId);
-            Assert.Equal("sessionId", output.SessionId);
-            Assert.Equal("replyTo", output.ReplyTo);
-            Assert.Equal("replyToSessionId", output.ReplyToSessionId);
             Assert.Equal("contentType", output.ContentType);
-            Assert.Equal("label", output.Subject);
-            Assert.Equal("to", output.To);
-            Assert.Equal("partitionKey", output.PartitionKey);
-            Assert.Equal("viaPartitionKey", output.TransactionPartitionKey);
-            Assert.Equal("deadLetterSource", output.DeadLetterSource);
-            Assert.Equal(1, output.EnqueuedSequenceNumber);
-            Assert.Equal(lockToken.ToString(), output.LockToken);
         }
 
-        private static ServiceBusReceivedMessage CreateReceivedMessage(Guid lockToken)
+        private static EventData CreateEventData()
         {
-            return ServiceBusModelFactory.ServiceBusReceivedMessage(
-                body: BinaryData.FromString("body"),
-                messageId: "messageId",
-                correlationId: "correlationId",
-                sessionId: "sessionId",
-                replyTo: "replyTo",
-                replyToSessionId: "replyToSessionId",
-                contentType: "contentType",
-                subject: "label",
-                to: "to",
-                partitionKey: "partitionKey",
-                viaPartitionKey: "viaPartitionKey",
-                deadLetterSource: "deadLetterSource",
-                enqueuedSequenceNumber: 1,
-                lockTokenGuid: lockToken);
+            return new EventData("body")
+            {
+                ContentType = "contentType",
+                CorrelationId = "correlationId",
+                MessageId = "messageId",
+            };
         }
 
-        private static BinaryData ConvertReceivedMessageToBinaryData(ServiceBusReceivedMessage message)
+        private static BinaryData ConvertEventDataToBinaryData(EventData @event)
         {
-            ReadOnlyMemory<byte> messageBytes = message.GetRawAmqpMessage().ToBytes().ToMemory();
-
-            byte[] lockTokenBytes = Guid.Parse(message.LockToken).ToByteArray();
-
-            // The lock token is a 16 byte GUID
-            const int lockTokenLength = 16;
-
-            byte[] combinedBytes = new byte[messageBytes.Length + lockTokenLength];
-
-            // The 16 lock token bytes go in the beginning
-            lockTokenBytes.CopyTo(combinedBytes.AsSpan());
-
-            // The AMQP message bytes go after the lock token bytes
-            messageBytes.CopyTo(combinedBytes.AsMemory(lockTokenLength));
-
-            return new BinaryData(combinedBytes);
+            return @event.GetRawAmqpMessage().ToBytes();
         }
     }
 }