+++ sdk/release_notes.md
@@ -8,9 +8,10 @@
 
 - <entry>
 
-### Microsoft.Azure.Functions.Worker.Sdk.Analyzers <version> (delete if not updated)
+### Microsoft.Azure.Functions.Worker.Sdk.Analyzers <version>
 
-- <entry>
+- Added an analyzer that will show a warning for types not supported by a binding attribute (#1505)
+- Added an analyzer that will suggest a code refactor for all of the types supported by a binding attribute (#1604)
 
 ### Microsoft.Azure.Functions.Worker.Sdk.Generators <version> (delete if not updated)